update: ESP32 compatibility

+++ src/CDisplayLine.h
@@ -5,7 +5,8 @@
 
 class CDisplayLine {
 public:
-  CDisplayLine(uint8_t x, uint8_t y, uint8 uiNoOfColumns, const uint8_t *pFont)
+  CDisplayLine(uint8_t x, uint8_t y, uint8_t uiNoOfColumns,
+               const uint8_t *pFont)
       : m_uiX(x), m_uiY(y), m_uiNoOfColumns(uiNoOfColumns), m_sLineToDraw(""),
         m_sLineDrawn(""), m_sLine(""), m_iScrollIdx(0), m_pFont(pFont) {}
 
@@ -22,7 +23,7 @@ public:
   void Scroll() {
     if (m_sLine.length() > m_uiNoOfColumns) {
       m_iScrollIdx++;
-      if (m_iScrollIdx > (int8)m_sLine.length())
+      if (m_iScrollIdx > (uint8_t)m_sLine.length())
         m_iScrollIdx = -m_uiNoOfColumns;
       std::string sEmptyLine(m_uiNoOfColumns, ' ');
       std::string sLineTmp = sEmptyLine + m_sLine + sEmptyLine;
@@ -32,11 +33,11 @@ public:
   }
   uint8_t m_uiX;
   uint8_t m_uiY;
-  uint8 m_uiNoOfColumns;
+  uint8_t m_uiNoOfColumns;
   std::string m_sLineToDraw;
   std::string m_sLineDrawn;
   std::string m_sLine;
-  int8 m_iScrollIdx;
+  int m_iScrollIdx;
   const uint8_t *m_pFont;
 };
